assembly/src/user-acceptance/requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-requests
psutil
pyautogui
opencv-python